pyproject.toml
[project]
name = "dino-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair<5",
    "dvc>=3.66.1",
    "hydra-core>=1.3.2",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "pytorch-lightning>=2.6.0",
    "streamlit>=1.19.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "wandb>=0.24.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "ipython>=9.9.0",
    "matplotlib>=3.10.8",
    "pip>=25.3",
]
